IBP-7142 move FieldEditorScreen.kt
