+++ shared/src/commonMain/kotlin/com/fieldbook/shared/screens/fields/FieldEditorScreen.kt
@@ -1,4 +1,4 @@
-package com.fieldbook.shared.screens
+package com.fieldbook.shared.screens.fields
 
 import androidx.compose.foundation.background
 import androidx.compose.foundation.border
@@ -20,6 +20,7 @@ import androidx.compose.material.icons.filled.ArrowBack
 import androidx.compose.material3.CircularProgressIndicator
 import androidx.compose.material3.ExperimentalMaterial3Api
 import androidx.compose.material3.FloatingActionButton
+import androidx.compose.material3.HorizontalDivider
 import androidx.compose.material3.Icon
 import androidx.compose.material3.IconButton
 import androidx.compose.material3.MaterialTheme
@@ -156,7 +157,7 @@ fun FieldEditorScreen(
                         LazyColumn(modifier = Modifier.fillMaxSize()) {
                             items(fieldsState.value!!) { field ->
                                 FieldListItem(field = field)
-                                androidx.compose.material3.HorizontalDivider()
+                                HorizontalDivider()
                             }
                         }
                     }
