+++ shared/src/iosMain/kotlin/com/fieldbook/shared/Main.kt
@@ -6,8 +6,8 @@ import androidx.compose.runtime.remember
 import androidx.compose.runtime.setValue
 import androidx.compose.ui.window.ComposeUIViewController
 import com.fieldbook.shared.screens.ConfigScreen
-import com.fieldbook.shared.screens.FieldEditorScreen
 import com.fieldbook.shared.screens.collect.CollectScreen
+import com.fieldbook.shared.screens.fields.FieldEditorScreen
 import com.fieldbook.shared.screens.preferences.PreferencesScreen
 import com.fieldbook.shared.screens.preferences.StorageDefinerScreen
 import com.fieldbook.shared.screens.preferences.StoragePreferencesScreen
